저장 안 한 거 추가

+++ make_train_data/Assets/script/FileInStream.cs
@@ -50,9 +50,9 @@ class FileInStream : MonoBehaviour
     {
         return JsonConvert.DeserializeObject<List<situation>>(GetFile_string(FileStream.Instance.json_Situation_filePath));
     }
-    public int GetFile_Award()
+    public float GetFile_Award()
     {
-        return JsonConvert.DeserializeObject<int>(GetFile_string(FileStream.Instance.json_Award_filePath));
+        return JsonConvert.DeserializeObject<float>(GetFile_string(FileStream.Instance.json_Award_filePath));
     }
     public string[,] GetFile_SignalType()
     {
